Enable default styles for array parameter delimiters
Styles are not required in OpenAPI, ensure that we use the default
delimiting styles when none is provided.

This change also slightly improves the error message when an unknown
parameter `style` is detected. Ideally that error should be raised
when the validator is instantiated, instead of when trying to use
parse the request.