test: Add Failing Test Case for Issue 535.
Add failing test case for
https://github.com/cdimascio/express-openapi-validator/issues/535.

The problem manifests itself when calling `middleware()` with an object
multiple times, as the object itself is mutated which might cause errors
on subsequent calls.

In this specific example, the parameters for GET /ping/{value} are
modified in such a way that a second call will cause a validation error.
This happens in `preprocessPathLevelParameters`.