feat: Add Allow Header on 405 (#560)
* chore: Fix Typos

* test: Add Test Case for Allow Header

The Allow header must be sent if the server responds with 405 to
indicate which request methods can be used.

See #467

* feat: Add Allow Header for 405 - Method Not Allowed

Adds Allow header to 405 error. This is required by RFC 7231
and indicates which request methods can be used.

Resolves #467

* doc: Adjust Error Handler in NestJS Example

When using a custom error handler like in the NestJS example, one needs
to set the headers on the response explicitly. The same is true when
using a custom express error handler.