Support ESM import of handler module in default resolver
Make the default request handler resolver work in ESM projects by
allowing a URL value for the handlersPath. If a URL is passed, then
a dynamic import will be used to load the handler module. If a string is
passed, then require() will be used.

Because this project uses moduleResolution:node in tsconfig, it's not
possible to write async import(...) directly in code. It has to be
obscured so that the compiler does not replace it, hence the HACK!
function that is essentially eval('import(...)').

ESM users should be aware that a file extension must be specified when
defining x-eov-operation-handler in schemas. For example:

```
{
  "get": {
    "summary": "Get list",
    "x-eov-operation-id": "getList",
    "x-eov-operation-handler": "list.js",
    "responses": {
       ...
```