fix: Dereference path parameters
The OpenAPI spec loader has a `discoverRoutes` method which explores an OpenAPI document
and gathers information about the paths and parameters used.
The list of discovered path parameters is used to install parameter-specific middleware in `src/openapi.validator.ts#installPathParams`
Path parameters declared with `$ref` were not detected in the `discoverRoutes` implementation, leading to the un-coerced values being used.
By dereferencing each path parameter when building this list, we should see the same behavior for referenced path parameters and for inline path parameters.

Closes https://github.com/cdimascio/express-openapi-validator/issues/803